// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/auditableEntityAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuditableEntityStatus } from './auditableEntityStatus';
import { BaseModelAC } from './baseModelAC';
import { EntityCategoryAC } from './entityCategoryAC';
import { EntityTypeAC } from './entityTypeAC';

export interface AuditableEntityAC extends BaseModelAC { 
    name?: string;
    status: AuditableEntityStatus;
    statusString?: string;
    description?: string;
    isStrategyAnalysisDone: boolean;
    version: number;
    newVersion: number;
    isNewVersion: boolean;
    selectedTypeId?: string;
    selectedTypeString?: string;
    selectedCategoryId?: string;
    selectedCategoryString?: string;
    engagementManagerStringList?: string;
    parentEntityId?: string;
    entityCategoryACList?: Array<EntityCategoryAC>;
    entityTypeACList?: Array<EntityTypeAC>;
}